Serparated userService and matchesService
